untangle `LocOffsets` from `Loc.h`
