+++ core/Loc.h
@@ -1,7 +1,8 @@
 #ifndef SORBET_AST_LOC_H
 #define SORBET_AST_LOC_H
 
-#include "Files.h"
+#include "core/Files.h"
+#include "core/LocOffsets.h"
 
 namespace sorbet::core {
 namespace serialize {
@@ -10,48 +11,6 @@ class SerializerImpl;
 class GlobalState;
 class Context;
 class MutableContext;
-
-constexpr int INVALID_POS_LOC = 0xfffffff;
-struct LocOffsets {
-    uint32_t beginLoc = INVALID_POS_LOC;
-    uint32_t endLoc = INVALID_POS_LOC;
-    uint32_t beginPos() const {
-        return beginLoc;
-    };
-
-    uint32_t endPos() const {
-        return endLoc;
-    }
-    bool exists() const {
-        return endLoc != INVALID_POS_LOC && beginLoc != INVALID_POS_LOC;
-    }
-    bool empty() const {
-        ENFORCE_NO_TIMER(exists());
-        return beginLoc == endLoc;
-    }
-    static LocOffsets none() {
-        return LocOffsets{INVALID_POS_LOC, INVALID_POS_LOC};
-    }
-    LocOffsets join(LocOffsets other) const;
-    // For a given LocOffsets, returns a zero-length version that starts at the same location.
-    LocOffsets copyWithZeroLength() const {
-        return LocOffsets{beginPos(), beginPos()};
-    }
-    // As above, but returns a zero-length version that starts at the end of the location.
-    LocOffsets copyEndWithZeroLength() const {
-        return LocOffsets{endPos(), endPos()};
-    }
-
-    std::string showRaw(const Context ctx) const;
-    std::string showRaw(const MutableContext ctx) const;
-    std::string showRaw(const GlobalState &gs, const FileRef file) const;
-    std::string showRaw() const;
-
-    bool operator==(const LocOffsets &rhs) const;
-
-    bool operator!=(const LocOffsets &rhs) const;
-};
-CheckSize(LocOffsets, 8, 4);
 
 class Loc final {
     struct {
